Add support for running provision command.
This commit adds support for running a command within a systemd-nspawn
container before building the PXE Live squashfs image. The use case for this is
to support running provisioning commands like puppet or ansible to provision
the image before building the PXE Live image. This is a WIP and interface is
subject to change.