Add support for custom CA.
To support booting live OS images when Grendel is configured with a custom CA,
we need to add this certificate to the trust store used be livenet during the
boot process (uses cURL). dracut will try to install the CA bundle
(/etc/ssl/certs/ca-bundle.crt) into the initramfs so that SSL certificate
checking will work properly. If running Grendel with a custom CA you can add
the CA certificate file to the following directory:

mkosi.extra/etc/pki/ca-trust/source/anchors/

The cert will then get included in the image and before we run dracut we
rebuild the trust store so the custom CA cert is included in the initramfs thus
allowing livenet to download the live image via trusted HTTPS.